pyproject.toml
[tool.poetry]
name = "agbee-api-python-example"
version = "0.1.0"
description = ""
authors = ["Hiroki Urase <doripu505@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
flet = "^0.22.1"
gql = {extras = ["all"], version = "^3.5.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
